+++ src/encoding/string_enc.c
@@ -68,7 +68,7 @@ String String_new(buffer_t *b, size_t n) {
 
     if (b) {
         int sizeNeeded = n;
-        if ( b->size - b->offset < sizeNeeded ) {
+        if ( (int)b->size - (int)b->offset < sizeNeeded ) {
             return init;
         }
         init.data.buffer.ptr = b->ptr;
